feat(ai-facade): 增强 AI 引擎回退机制和错误处理
- 新增 uncaughtFailureEvent 参数确保下游总能收到一致的终端事件形态
- 添加 downstreamCollectorFailed 标志防止 Flow 透明度异常
- 实现合成失败事件机制处理未捕获的异常情况
- 为 DefaultAiChat 和 DefaultAiImage 配置失败事件构造函数
- 优化 SdWebUiImageEngine 的日志记录和进度轮询异常处理
- 修复 Text2ImagePayload 中 hrScale 类型从 Int 到 Float 的转换
- 增强 ErrorMapper 中的取消异常处理逻辑
- 优化 JSON 解析安全性避免非对象字面量导致的崩溃
- 为 SdWebUi 模型类添加必要的序列化注解支持
- 改进测试工具 RecordingClient 的线程安全性

+++ feature/sdwebui/src/commonMain/kotlin/ciyin/sdwebui/payload/Text2ImagePayload.kt
@@ -40,7 +40,7 @@ data class Text2ImagePayload(
     @SerialName("enable_hr") val enableHr: Boolean,
     @SerialName("firstphase_width") val firstphaseWidth: Int,
     @SerialName("firstphase_height") val firstphaseHeight: Int,
-    @SerialName("hr_scale") val hrScale: Int,
+    @SerialName("hr_scale") val hrScale: Float,
     @SerialName("hr_upscaler") val hrUpscaler: String,
     @SerialName("hr_second_pass_steps") val hrSecondPassSteps: Int,
     @SerialName("hr_resize_x") val hrResizeX: Int,
